+++ dfetch/commands/init.py
@@ -10,6 +10,7 @@ import shutil
 
 import dfetch.commands.command
 from dfetch.resources import TEMPLATE_PATH
+from dfetch import DEFAULT_MANIFEST_NAME
 
 logger = logging.getLogger(__name__)
 
@@ -20,8 +21,6 @@ class Init(dfetch.commands.command.Command):
     Generate a manifest that can be used as basis for a project.
     """
 
-    MANIFEST_NAME = "manifest.yaml"
-
     @staticmethod
     def create_menu(subparsers: "argparse._SubParsersAction") -> None:
         """Add the parser menu for this action."""
@@ -31,11 +30,11 @@ class Init(dfetch.commands.command.Command):
         """Perform the init."""
         del args  # unused
 
-        if os.path.isfile(self.MANIFEST_NAME):
-            logger.warning(f"{self.MANIFEST_NAME} already exists!")
+        if os.path.isfile(DEFAULT_MANIFEST_NAME):
+            logger.warning(f"{DEFAULT_MANIFEST_NAME} already exists!")
             return
 
         with TEMPLATE_PATH as template_path:
-            dest = shutil.copyfile(template_path, self.MANIFEST_NAME)
+            dest = shutil.copyfile(template_path, DEFAULT_MANIFEST_NAME)
 
         logger.info(f"Created {dest}")