+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Release 0.0.5 (In development)
 * Fix ``dfetch import`` command.
 * Improve template.
 * If no ``dst`` is given for a project, use name of project instead.
+* Fixes #28: Rename manifest.yaml to dfetch.yaml
 
 Release 0.0.4 (released 2020-11-12)
 ===================================